Disallow copy/move operations for most classes
Most internal classes are intended to be passed around by pointer or
reference, so copy or move operations are likely to be a mistake.

I've thought about using more of a RAII approach, but I don't think it
would work well for the AST objects. They're so nested that copy or move
operations would be very expensive. Perhaps we should support it for the
root node (FileNode). Something to possibly revist another day.