Manage FileContent using a smart pointer
I'm planning to convert more of the codebase to use smart pointers.
Probably not going to go all-in on using them everywhere, but I'd like to
switch to using unique_ptr where the object is actually owned. Everywhere
else, I think I'm going to continue using raw pointers, at least for now.

The ownership model for data in MOT is fairly straight-forward, so
shared_ptr is less necessary. That being said, I wouldn't be surprised if
I eventually switch to it anyway.